Extraigo la funcionalidad de generación de indicadores
